requirements.txt
# === KodiBOT Backend Dependencies ===

# 🚀 Web Framework & API
fastapi==0.115.6                    # Modern API framework with validation
uvicorn[standard]==0.34.0           # ASGI server for FastAPI

# 🗄️ Database & ORM  
sqlalchemy==2.0.36                  # SQL toolkit and ORM

# 🧠 AI & Machine Learning
openai==1.58.1                      # OpenAI GPT API client

# 🔧 Configuration & Environment
python-dotenv==1.0.1                # Environment variables (.env files)

# 📋 Data Validation
pydantic==2.10.5                    # Data validation using type hints

# 🌐 HTTP Requests (for tests)
requests==2.32.3                    # HTTP library for test_chat.py
